eval: abstract provider
- move CodeGenerator to provider module
- dynamically import PROVIDER module
- resolve model choices from provider module
- add google provider
