+++ critic.py
@@ -1,15 +1,6 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import dataclasses
-import json
-from typing import Optional
-
-try:
-    import openai  # type: ignore
-except ModuleNotFoundError:
-    openai = None
-
-from dslx_text import strip_fences
 
 
 @dataclasses.dataclass
@@ -90,89 +81,3 @@ def load_dslx_critic_reference(prompt_md_path: str) -> str:
     if for_loops:
         parts.append(for_loops)
     return "\n\n".join(parts).strip()
-
-
-def _chat_kwargs(model: str, reasoning_effort: Optional[str], messages, need_reasoning_effort: set[str]):
-    if model in need_reasoning_effort:
-        assert reasoning_effort is not None
-        return {'model': model, 'reasoning_effort': reasoning_effort, 'messages': messages}
-    return {'model': model, 'messages': messages}
-
-
-def _parse_critic_json(text: str) -> dict:
-    # Allow the critic to wrap the JSON in a triple-backtick block.
-    raw = strip_fences(text).strip()
-    return json.loads(raw)
-
-
-def run_critic(
-    *,
-    critic_model: str,
-    critic_reasoning_effort: Optional[str],
-    need_reasoning_effort: set[str],
-    dslx_critic_reference: str,
-    prompt: str,
-    signature: str,
-    requirements: str,
-    generated_code: str,
-) -> CriticResult:
-    if openai is None:
-        raise RuntimeError(
-            'The "openai" Python package is required to run evaluations. '
-            'Install it (e.g. `pip install -r requirements.txt`) and retry.'
-        )
-
-    candidate = strip_fences(generated_code).strip()
-    user_message = (
-        "Problem prompt:\n"
-        f"{prompt}\n\n"
-        "Signature:\n"
-        f"{signature}\n\n"
-        "Requirements:\n"
-        f"{requirements}\n\n"
-        f"{dslx_critic_reference}\n\n"
-        "Candidate DSLX implementation:\n"
-        "```dslx\n"
-        f"{candidate}\n"
-        "```"
-    )
-
-    client = openai.Client()
-    messages = [
-        {"role": "system", "content": CRITIC_SYSTEM_PROMPT},
-        {"role": "user", "content": user_message},
-    ]
-
-    last_text = ""
-    for _attempt in range(1, 3):
-        response = client.chat.completions.create(
-            **_chat_kwargs(critic_model, critic_reasoning_effort, messages, need_reasoning_effort)
-        )
-        last_text = response.choices[0].message.content.strip()
-        try:
-            parsed = _parse_critic_json(last_text)
-        except Exception as e:
-            # Ask for a regeneration with strictly valid JSON.
-            messages.append({"role": "assistant", "content": last_text})
-            messages.append({
-                "role": "user",
-                "content": (
-                    "Your previous response was not valid JSON and could not be parsed. "
-                    f"Parsing error: {e}. "
-                    "Please respond again with ONLY valid JSON matching the schema exactly."
-                ),
-            })
-            continue
-
-        ok = bool(parsed.get("pass"))
-        confidence = float(parsed.get("confidence", 0.0))
-        message = str(parsed.get("message", "")).strip()
-        raw_json = strip_fences(last_text).strip()
-        return CriticResult(ok=ok, confidence=confidence, message=message, raw_json=raw_json)
-
-    return CriticResult(
-        ok=False,
-        confidence=0.0,
-        message="Critic did not return valid JSON after retries.",
-        raw_json=strip_fences(last_text).strip(),
-    )
